client.utils: Include the content of base_utils inside utils.
In order to make the API documentation simpler to understand,
move base_utils functions to utils and make the necessary arrangements.
